refactored class breakout
