+++ intellij-projects/java-spring-refactor/src/main/java/ai/codium/refactor/config/Config.java
@@ -1,0 +1,15 @@
+package ai.codium.refactor.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.core.JdbcTemplate;
+
+import javax.sql.DataSource;
+
+@Configuration
+public class Config {
+    @Bean
+    public JdbcTemplate jdbcTemplate(DataSource dataSource) {
+        return new JdbcTemplate(dataSource);
+    }
+}
