+++ intellij-projects/java-spring-refactor/src/main/java/ai/codium/refactor/entity/UserProfile.java
@@ -1,0 +1,28 @@
+package ai.codium.refactor.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+
+@Entity
+public class UserProfile {
+
+    @Id
+    private Long id;
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+}
