+++ intellij-projects/java-spring-refactor/src/main/java/ai/codium/refactor/repository/ProfileRepo.java
@@ -1,0 +1,10 @@
+package ai.codium.refactor.repository;
+
+import ai.codium.refactor.entity.UserProfile;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProfileRepo extends JpaRepository<UserProfile, Long> {
+    UserProfile findUserProfileById(Long id);
+}
